# agenx-drawio-app/src/main/resources/static/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "API Documentation",
    "version" : "1.0"
  },
  "paths" : {
...
    "/api/v1/create_session" : {
      "get" : {
        "summary" : "createSession",
        "description" : "createSession",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "agentId" : {
                    "type" : "string",
                    "description" : "智能体ID"
                  },
                  "userId" : {
                    "type" : "string",
                    "description" : "用户ID"
                  }
                },
                "description" : ""
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "info" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "data" : {
                      "type" : "object",
                      "properties" : {
                        "sessionId" : {
                          "type" : "string",
                          "description" : "会话ID"
                        }
                      },
                      "description" : ""
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat" : {
      "post" : {
        "summary" : "chat",
        "description" : "chat",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "agentId" : {
                    "type" : "string",
                    "description" : "智能体ID"
                  },
                  "userId" : {
                    "type" : "string",
                    "description" : "用户ID"
                  },
                  "sessionId" : {
                    "type" : "string",
                    "description" : "会话ID"
                  },
                  "message" : {
                    "type" : "string",
                    "description" : "消息内容"
                  }
                },
                "description" : ""
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "info" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "data" : {
                      "type" : "object",
                      "properties" : {
                        "content" : {
                          "type" : "string",
                          "description" : "回复内容"
                        }
                      },
                      "description" : ""
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}